clean up warnings from msvc

+++ src/learner/evaluation-inl.hpp
@@ -132,7 +132,7 @@ struct EvalAMS : public IEvaluator {
         b_fp += wt;
       }
       if (rec[i].first != rec[i+1].first) {
-        double ams = sqrtf(2*((s_tp+b_fp+br) * log(1.0 + s_tp/(b_fp+br)) - s_tp));
+        double ams = sqrt(2*((s_tp+b_fp+br) * log(1.0 + s_tp/(b_fp+br)) - s_tp));
         if (tams < ams) {
           thresindex = i;
           tams = ams;
@@ -141,9 +141,9 @@ struct EvalAMS : public IEvaluator {
     }
     if (ntop == ndata) {
       fprintf(stderr, "\tams-ratio=%g", static_cast<float>(thresindex) / ndata);
-      return tams;
+      return static_cast<float>(tams);
     } else {
-      return sqrtf(2*((s_tp+b_fp+br) * log(1.0 + s_tp/(b_fp+br)) - s_tp));
+      return static_cast<float>(sqrt(2*((s_tp+b_fp+br) * log(1.0 + s_tp/(b_fp+br)) - s_tp)));
     }
   }
   virtual const char *Name(void) const {
@@ -171,7 +171,7 @@ struct EvalPrecisionRatio : public IEvaluator{
     utils::Assert(preds.size() == info.labels.size(), "label size predict size not match");
     std::vector< std::pair<float, unsigned> > rec;
     for (size_t j = 0; j < preds.size(); ++j) {
-      rec.push_back(std::make_pair(preds[j], j));
+      rec.push_back(std::make_pair(preds[j], static_cast<unsigned>(j)));
     }
     std::sort(rec.begin(), rec.end(), CmpFirst);
     double pratio = CalcPRatio(rec, info);
@@ -207,7 +207,7 @@ struct EvalAuc : public IEvaluator {
   virtual float Eval(const std::vector<float> &preds,
                      const MetaInfo &info) const {
     utils::Check(preds.size() == info.labels.size(), "label size predict size not match");
-    std::vector<unsigned> tgptr(2, 0); tgptr[1] = preds.size();
+    std::vector<unsigned> tgptr(2, 0); tgptr[1] = static_cast<unsigned>(preds.size());
     const std::vector<unsigned> &gptr = info.group_ptr.size() == 0 ? tgptr : info.group_ptr;
     utils::Check(gptr.back() == preds.size(),
                  "EvalAuc: group structure must match number of prediction");
@@ -264,7 +264,7 @@ struct EvalRankList : public IEvaluator {
     utils::Check(preds.size() == info.labels.size(),
                   "label size predict size not match");
     // quick consistency when group is not available
-    std::vector<unsigned> tgptr(2, 0); tgptr[1] = preds.size();
+    std::vector<unsigned> tgptr(2, 0); tgptr[1] = static_cast<unsigned>(preds.size());
     const std::vector<unsigned> &gptr = info.group_ptr.size() == 0 ? tgptr : info.group_ptr;
     utils::Assert(gptr.size() != 0, "must specify group when constructing rank file");
     utils::Assert(gptr.back() == preds.size(),
@@ -339,7 +339,7 @@ struct EvalNDCG : public EvalRankList{
     for (size_t i = 0; i < rec.size() && i < this->topn_; ++i) {
       const unsigned rel = rec[i].second;
       if (rel != 0) { 
-        sumdcg += ((1 << rel) - 1) / logf(i + 2);
+        sumdcg += ((1 << rel) - 1) / log(i + 2.0);
       }
     }
     return static_cast<float>(sumdcg);
